Further work on tasks + corrected simulateExt. (#16)
* Some changes fell through in pull request.
* Checking for the `Pending` status of runnables was done in the wrong place.

The downside is that `hear` now needs a process map parameter. Tasks can still be scheduled when they are not ready (i.e. prior to the runnable at the head of their execution queue being tagged as Pending.). Tasks need to be triggered by the same labels that makes runnables `Pending` (i.e. `TICK`) to go from Inactive.